+++ tilebench/middleware.py
@@ -39,7 +39,10 @@ class VSIStatsMiddleware(BaseHTTPMiddleware):
         handler.close()
 
         if rio_stream or curl_stream:
-            results = analyse_logs(rio_stream, curl_stream,)
+            rio_lines = rio_stream.getvalue().splitlines()
+            curl_lines = curl_stream.read().splitlines()
+
+            results = analyse_logs(rio_lines, curl_lines)
             head_results = "head;count={count}".format(**results["HEAD"])
             list_results = "list;count={count}".format(**results["LIST"])
             get_results = "get;count={count};size={bytes}".format(**results["GET"])